requirements.txt
streamlit>=1.31.0
httpx>=0.26.0
pandas>=2.2.0
rapidfuzz>=3.6.0
transformers>=4.37.0
torch>=2.2.0
gspread>=6.0.0
oauth2client>=4.1.3
python-dotenv>=1.0.1
ebaysdk>=2.1.5
